Refactor text effect file structure
